requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
python-dotenv
passlib